WIP: save work before merging main

+++ src/application/use-cases/order/delete-order.use-case.ts
@@ -3,6 +3,7 @@ import { IUseCase } from '../base.use-case';
 import { Orders } from 'src/domain/entities/orders.entity';
 import { IOrdersRepository } from 'src/domain/repositories/order.repository.interface';
 import { ForbiddenError } from 'src/application/errors/forbidden.error';
+import { NotFoundError } from 'src/application/errors/not-found.error';
 
 export class DeleteOrderUseCase implements IUseCase<
   DeleteOrderCommand,
@@ -13,6 +14,7 @@ export class DeleteOrderUseCase implements IUseCase<
   async execute(command: DeleteOrderCommand): Promise<Orders> {
     const { orderId } = command;
     const order = await this.orderRepository.findById(orderId);
+    if (!order) throw new NotFoundError('Order not found');
     if (order.userId !== orderId) throw new ForbiddenError();
     return this.orderRepository.deleteOrder(orderId);
   }
